Move test to test/api/from-outside.js
